+++ net/levin/src/bucket_stream.rs
@@ -1,13 +1,15 @@
-use std::marker::PhantomData;
+//! This module provides a `BucketStream` struct, which is a stream of `Bucket`s, 
+//! where only the header is decoded. If you are a user of this library you should 
+//! probably use `MessageStream` instead.
+
+
+
 use std::task::Poll;
 
 use bytes::{Buf, BytesMut};
 use futures::stream::Stream;
 use futures::{ready, AsyncRead};
 use pin_project::pin_project;
-
-use crate::header::Flags;
-use crate::{LEVIN_SIGNATURE, PROTOCOL_VERSION};
 
 use super::{Bucket, BucketError, BucketHead};
 
@@ -131,73 +133,6 @@ impl<S: AsyncRead + std::marker::Unpin> Stream for BucketStream<S> {
                         continue;
                     }
                 }
-            }
-        }
-    }
-}
-
-pub trait MessageDecoder {
-    type Message;
-    type Error: From<BucketError>;
-
-    fn decode_message(
-        buf: &[u8],
-        flags: Flags,
-        have_to_return: bool,
-        command: u32,
-    ) -> Result<Self::Message, Self::Error>;
-}
-
-#[pin_project]
-pub struct MessageStream<D: MessageDecoder, S: AsyncRead + std::marker::Unpin> {
-    #[pin]
-    bucket_stream: BucketStream<S>,
-    phantom: PhantomData<D>,
-}
-
-impl<D: MessageDecoder, S: AsyncRead + std::marker::Unpin> MessageStream<D, S> {
-    pub fn new(stream: S) -> Self {
-        MessageStream { 
-            bucket_stream: BucketStream::new(stream), 
-            phantom: PhantomData
-        }
-    }
-} 
-
-impl<D: MessageDecoder, S: AsyncRead + std::marker::Unpin> Stream for MessageStream<D, S> {
-    type Item = Result<D::Message, D::Error>;
-
-    fn poll_next(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> Poll<Option<Self::Item>> {
-        let this = self.project();
-        match ready!(this.bucket_stream.poll_next(cx)).expect("BucketStream will never return None")
-        {
-            Err(e) => Poll::Ready(Some(Err(e.into()))),
-            Ok(bucket) => {
-                if bucket.header.signature != LEVIN_SIGNATURE {
-                    return Err(BucketError::IncorrectSignature(bucket.header.signature))?;
-                }
-
-                if bucket.header.protocol_version != PROTOCOL_VERSION {
-                    return Err(BucketError::UnknownProtocolVersion(
-                        bucket.header.protocol_version,
-                    ))?;
-                }
-
-                if bucket.header.return_code < 0
-                    || (bucket.header.return_code == 0 && bucket.header.flags == Flags::RESPONSE)
-                {
-                    return Err(BucketError::Error(bucket.header.return_code))?;
-                }
-
-                Poll::Ready(Some(D::decode_message(
-                    &bucket.body,
-                    bucket.header.flags,
-                    bucket.header.have_to_return_data,
-                    bucket.header.command,
-                )))
             }
         }
     }
